restructure tests, more fixing
