+++ Jint.Tests.Test262/BuiltIns/DateTests.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit;
 
-namespace Jint.Tests.Test262
+namespace Jint.Tests.Test262.BuiltIns
 {
     public class DateTests : Test262Test
     {
