+++ Jint.Tests.Test262/BuiltIns/ReflectTests.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit;
 
-namespace Jint.Tests.Test262
+namespace Jint.Tests.Test262.BuiltIns
 {
     public class ReflectTests : Test262Test
     {
