+++ Jint.Tests.Test262/BuiltIns/SetTests.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit;
 
-namespace Jint.Tests.Test262
+namespace Jint.Tests.Test262.BuiltIns
 {
     public class SetTests : Test262Test
     {
