+++ Jint.Tests.Test262/BuiltIns/SymbolTests.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit;
 
-namespace Jint.Tests.Test262
+namespace Jint.Tests.Test262.BuiltIns
 {
     public class SymbolTests : Test262Test
     {
